Adding the creation of a new event in EventController

+++ BookingEvent/src/main/java/com/example/BookingEvent/Models/EventRequest.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 @Setter
 
 public class EventRequest {
-    private int eventID;
+    private long eventID;
     private String eventName;
     private LocalDate eventDate;
     private String  eventLocation;
